add test for MathCalcul and StringConvertion TG-35 TG-34 TG-39
